==> 4 Sensitivity.Rmd <==
---
title: "Parameter sensitivity analysis"
output: html_notebook
---


Loading R packages

```{r}
library(SWATplusR)
library(sensitivity)
library(fast)
library(hydroGOF)
library(dplyr)
library(forcats)
library(tidyr)
library(ggplot2)
library(lhs)
library(purrr)
library(lubridate)
library(magrittr)
library(reshape2)

```

#Loading model

```{r}
# The path where the SWAT project is written
path_2012 <- "C:/Veracruz/SWAT/github-rswat-ver/swat_models/swat_t_canmx"

```

#Import observed files

```{r}

#Create a list of summary data for pixquic, contains dataframes
gav <- list()
#Add observed discharge data
gav$obs <- read.csv("C:/Veracruz/SWAT/github-rswat-ver/data/q_obs_gav.csv")%>%
  mutate(Date = as.Date(Date))


#Create a list of summary data for pixquic, contains dataframes
pix <- list()
#Add observed discharge data
pix$obs <- read.csv("C:/Veracruz/SWAT/github-rswat-ver/data/q_obs_pix.csv")%>%
  mutate(Date = as.Date(Date))
```

#Run the manual calibration run

```{r}

q_sim_man <- run_swat2012(project_path = path_2012,
                           output = define_output(file = "rch",
                                                  variable = "FLOW_OUT",
                                                  unit = c(6, 15)))
#Plot gavilanes
q_sim_man%>%
  subset(date %in% gav$obs$Date,)%>%
  ggplot() + geom_line(aes(date,FLOW_OUT_15)) + geom_line(data=gav$obs, aes(Date,Q_m3_s), color = "blue")

#print statistics of fit
q_sim_man%>%
  subset(date %in% gav$obs$Date,)%$%
  print(c("R^2 = ", (cor(FLOW_OUT_15, gav$obs$Q_m3_s))^2,
          "NSE = ", NSE(FLOW_OUT_15, gav$obs$Q_m3_s),
          "KGE = ", KGE(FLOW_OUT_15, gav$obs$Q_m3_s),
          "PBIAS = ", pbias(FLOW_OUT_15, gav$obs$Q_m3_s)))

#Plot Pixquiac
q_sim_man%>%
  subset(date %in% pix$obs$Date,)%>%
  ggplot() + geom_line(aes(date,FLOW_OUT_6)) + geom_line(data=pix$obs, aes(Date,Q_m3_s), color = "blue")+
  geom_line(data=pix$obs, aes(Date, Q_m3_s + 0.56), color = "red")


#print statistics of fit
q_sim_man%>%
  subset(date %in% pix$obs$Date,)%$%
  print(c("R^2 = ", (cor(FLOW_OUT_6, pix$obs$Q_m3_s))^2,
          "NSE = ", NSE(FLOW_OUT_6, pix$obs$Q_m3_s),
          "KGE = ", KGE(FLOW_OUT_6, pix$obs$Q_m3_s),
          "PBIAS = ", pbias(FLOW_OUT_6, pix$obs$Q_m3_s)))

```

# Type of change
r = ‘relchg’, multiply
v = “absval”, replace
a = “abschg”, add 


Visualization of uncertainty Gavilanes watershed

```{r}

#Create a theme
My_Theme = theme(plot.title = element_text(face = "bold", size= 16),axis.title = element_text(size= 14), 
                 axis.text = element_text(size = 18), legend.title = element_text(size=16), 
                 legend.text = element_text(size=16)) + theme_classic(base_size = 16)

#Plot timeseries gavilanes
ggplot() + geom_ribbon(data=gav$obs, aes(x =Date, ymin = q5, ymax = q95),fill = "green", alpha = .5) +
  geom_line(data=gav$obs, aes(Date,Q_m3_s, colour="obs")) +
  geom_line(data=gav$sim_pwn, aes(date,run_0563, colour="best sim"))+
  theme(legend.position="top") +labs(colour="",x=" Date", y="Discharge m^3/S") + My_Theme


#Create flow duration curves
gav$fdc <- gav$obs%>%
  arrange(., desc(Q_m3_s))%>%
  mutate(q5 = sort(gav$obs$q5,decreasing = T))%>%
  mutate(q95 = sort(gav$obs$q95,decreasing = T))%>%
  mutate(p = seq(0,1, length.out = nrow(.))) # This needs to consider n + 1 


# Plot FDC
ggplot() + 
  geom_ribbon(data=gav$fdc, aes(x =p, ymin = q5, ymax = q95, colour = "5-95 PPU" ), fill = "green", alpha = .5) +
  geom_line(data=gav$fdc, aes(p, Q_m3_s, colour="obs"))+ geom_polygon() +
  theme(legend.position="top") +labs(colour="",x=" Prob of exceedance", y="Discharge m^3/S") +
  scale_y_continuous(trans='log2') + My_Theme


# Plot Best sim and 1 to 1 line
plot(gav$obs$Q_m3_s,gav$sim_pwn$run_0563) 
abline( a= 0, b = 1 , col="blue")


print(c("R^2 = ", (cor(gav$sim_pwn$run_0563, gav$obs$Q_m3_s))^2,
          "NSE = ", NSE(gav$sim_pwn$run_0563, gav$obs$Q_m3_s),
          "KGE = ", KGE(gav$sim_pwn$run_0563, gav$obs$Q_m3_s),
          "PBIAS = ", pbias(gav$sim_pwn$run_0563, gav$obs$Q_m3_s)))

```

#Fast sensitivity analysis

```{r}
#Define the parameter names

par_names <- c("SOL_AWC.sol|change = relchg",
               "SOL_K.sol|change = relchg",
               "SOL_Z.sol|change = relchg",
               "CN2.mgt|change = abschg",
               "ALPHA_BF.gw|change = absval",
               "GW_DELAY.gw|change = absval",
               "GW_REVAP.gw|change = absval",
               "GWQMN.gw|change = absval",
               "RCHRG_DP.gw|change = absval",
               "REVAPMN.gw|change = absval",
               "CH_K2.rte|change = absval",
               "CH_N2.rte|change = absval",
               "EPCO.hru|change = absval",
               "ESCO.hru|change = absval",
               "HRU_SLP.hru|change = relchg",
               "SLSUBBSN.hru|change = relchg",
               "SURLAG.bsn|change = absval",
               "CANMX.hru|change = relchg",
               "GW_SPYLD.gw|change = absval", #New parameters
               "ALPHA_BNK.rte|change = absval",
               "LAT_TTIME.hru|change = absval",
               "SLSOIL.hru|change = absval",
               "DEP_IMP.hru|change = absval"
               )

#Parameter sampling with the FAST method

par_fast <- fast_parameters(
  minimum = c(-0.35, -0.5, -0.3, -15, 0, 0, 0.02, 0, 0, 0, 5, 0, 0, 0, -0.2, -0.3, 0,-0.3,0, 0, 0, 0, 0),
  maximum = c(0.35, 0.5, 0.3, 15, 1, 650, 0.2, 2000, 1, 500, 130, 0.3, 1, 1, 0.2, 0.3, 10, 0.3, 0.4, 1, 180, 150, 6000),
  names = par_names) %>%
  as_tibble()

#Model simulation and evaluation
q_fast <- run_swat2012(project_path = path_2012,
            output = list(q_out = define_output(file = "rch",
                                  variable = "FLOW_OUT",
                                  unit = c(6,15))), # gavilanes = 15, pixquiac = 6
            parameter = par_fast,
            start_date = "2014-01-01",
            end_date = "2017-12-31",
            years_skip = 1,
            n_thread = 4)

```

Fast sensitivity Gavilanes

```{r}
#Gavilanes

#Adding results from fast to Gavilanes
gav$sim_fast <- q_fast$simulation$q_out_15[q_fast$simulation$q_out_15$date %in% gav$obs$Date,]


#Nash Sutclife
gav$nse_fast <- gav$sim_fast %>%
  select(-date) %>%
  map_dbl(., ~NSE(.x, gav$obs$Q_m3_s))

#Coefficient of determination

gav$r2_fast <- gav$sim_fast%>%
  select(-date)%>%
  map_dbl(., ~cor(., gav$obs$Q_m3_s, method="pearson")^2)

# Kling-Gupta efficiency criterion (KGE)

gav$kge_fast <- gav$sim_fast%>%
  select(-date)%>%
  map_dbl(., ~KGE(., gav$obs$Q_m3_s))


#Sensitivity using different statistics of fit
q_fast$sens.nse.gav <- sensitivity(gav$nse_fast, 23) # These are the number of parameters
q_fast$sens.r2.gav <- sensitivity(gav$r2_fast, 23) # These are the number of parameters
q_fast$sens.kge.gav <- sensitivity(gav$kge_fast, 23) # These are the number of parameters

#higher senstitivities change depending on the statistics of fit
q_fast$result.gav <- tibble(parameter = q_fast$parameter$definition$par_name,
                      fast      = q_fast$sens.kge.gav) %>%
  mutate(parameter = factor(parameter) %>% fct_reorder(., fast))


ggplot(data = q_fast$result.gav) +
  geom_bar(aes(x = parameter, y = fast), stat = "identity") +
  xlab("Parameter") +
  ylab("Sensitivity") +
  coord_flip() +
  theme_bw()+
  theme(text = element_text(size = 14))


```
